Move InputOutput in component package
